Move Factory class out of the Extractor file
