+++ lib/Extractor.py
@@ -26,15 +26,3 @@ class CsvExtractor(Extractor):
             .option("inferSchema", "true")
             .load(path)
         )
-
-
-class Factory:
-
-    def __init__(self, conf):
-        self.conf = conf
-
-    def make_extractor(self, spark):
-        if self.conf["enable.hive"] == "true":
-            return HiveExtractor(spark)
-
-        return CsvExtractor(spark)
